Refactor JwtUtility class

+++ FamilijaApi/Controllers/AddressesController.cs
@@ -6,6 +6,12 @@ using FamilijaApi.DTOs;
 using FamilijaApi.Models;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Authorization;
+using FamilijaApi.Utility;
+using FamilijaApi.DTOs.Requests;
+using FamilijaApi.Configuration;
+using Microsoft.Extensions.Options;
+using FamilijaApi.Data;
+using Microsoft.IdentityModel.Tokens;
 
 namespace FamilijaApi.Controllers
 {
@@ -16,19 +22,30 @@ namespace FamilijaApi.Controllers
     {
         private IAdddressesRepo _addressRepo;
         private IMapper _mapper;
+        private readonly IJwtUtil _jwtTokenUtil;
 
-        public AddressesController (IAdddressesRepo adddressesRepo, IMapper mapper)
+        public AddressesController (IRoleRepo roleRepo, IAuthRepo authRepo, IUserRepo userRepo, IAdddressesRepo adddressesRepo, IMapper mapper, IOptionsMonitor<Jwtconfig> optionsMonitor, TokenValidationParameters tokenValidation)
         {
             _addressRepo = adddressesRepo;
             _mapper = mapper;
+            _jwtTokenUtil=new JwtTokenUtility(authRepo, userRepo, roleRepo, optionsMonitor.CurrentValue, tokenValidation);
         }
         
-        [HttpGet("{id}")]
-        public async Task<ActionResult<AddressReadDto>> GetAddress(int id)
+        [HttpGet]
+        public async Task<IActionResult> GetAddress([FromBody] TokenRequest token)
         {
-            var content = await _addressRepo.GetAddress(id);
-            if (content == null) return NoContent();
-            return Ok(_mapper.Map<AddressReadDto>(content));
+            var auth=await _jwtTokenUtil.VerifyAndGenerateToken(token);
+            var finalAuth= _mapper.Map<AuthResult>(auth);
+            if(auth.Success){            
+                var content = await _addressRepo.GetAddress(auth.Id);
+                if (content == null) 
+                    return NotFound(finalAuth);
+                return Ok(new CommunicationModel<AddressReadDto>(){
+                    GenericModel=_mapper.Map<AddressReadDto>(content),
+                    AuthResult=finalAuth
+                });
+            }
+            return Unauthorized(finalAuth);
         }
 
         //[HttpPost]
@@ -41,34 +58,47 @@ namespace FamilijaApi.Controllers
         //    return Created("", addres);
         //}
 
-        [HttpPut("{id}")]
-        public async Task<IActionResult> UpdateAddress(int id, AddressUpdateDto addressUpdateDto)
+        [HttpPut]
+        public async Task<IActionResult> UpdateAddress([FromBody] GetCommunicationModel<AddressUpdateDto> data)
         {
-            var updateModelAddress = _addressRepo.GetAddress(id).Result;
-            if (updateModelAddress == null)
-            {
-                return NotFound();
+            var auth=await _jwtTokenUtil.VerifyAndGenerateToken(data.TokenRequest);
+            var finalAuth= _mapper.Map<AuthResult>(auth);
+            if(auth.Success){ 
+                var updateModelAddress = _addressRepo.GetAddress(auth.Id).Result;
+                if (updateModelAddress == null)
+                {
+                    finalAuth.Errors.Add("Addres is not found");
+                    return NotFound(finalAuth);
+                }
+                _mapper.Map(data.GenericModel, updateModelAddress);
+                _addressRepo.UpdateAddress(updateModelAddress);
+                await _addressRepo.SaveChanges();
+                return Ok(finalAuth);
             }
+            finalAuth.Errors.Add("Unauthorized");
+            return Unauthorized();
 
-            _mapper.Map(addressUpdateDto, updateModelAddress);
-            _addressRepo.UpdateAddress(updateModelAddress);
-            await _addressRepo.SaveChanges();
-            return Ok();
 
         }
 
-        [HttpDelete("{id}")]
-        public ActionResult DeleteAddress(int id)
+        [HttpDelete]
+        public async Task<ActionResult> DeleteAddress([FromBody] TokenRequest token)
         {
-            var deleteModelAddress = _addressRepo.GetAddress(id).Result;
-            if (deleteModelAddress == null)
-            {
-                return NotFound();
+            var auth=await _jwtTokenUtil.VerifyAndGenerateToken(token);
+            var finalAuth= _mapper.Map<AuthResult>(auth);
+            if(auth.Success){
+                var deleteModelAddress = _addressRepo.GetAddress(auth.Id).Result;
+                if (deleteModelAddress == null)
+                {
+                    finalAuth.Errors.Add("User not found");
+                    return NotFound(finalAuth);
+                }
+                _addressRepo.DeleteAdress(deleteModelAddress);
+                await _addressRepo.SaveChanges();
+                return NoContent();
             }
-
-            _addressRepo.DeleteAdress(deleteModelAddress);
-            _addressRepo.SaveChanges();
-            return NoContent();
+            finalAuth.Errors.Add("Unauthorized");
+            return Unauthorized();
         }
     }
 }
